completed Chapter 13 and this book!
This was a great realisation and starter for functional programming! As there was not much exercises, I would like to take this knowledge to practice. Next plan of action is 2 forks => ReactJS+Ramda, however Ramda does not support Containers only in build JS structures; Test waters with FULL FP LANGUAGE (SCALA). Tending towards this more currently...